[#6181] Split up test_catalog into separate suites
