+++ scripts/core_tests_list.json
@@ -7,12 +7,16 @@
     "test_dynamic_peps",
     "test_iadmin",
     "test_ibun",
+    "test_icd",
     "test_ichksum",
     "test_ichmod",
     "test_icommands_file_operations",
     "test_icp",
+    "test_iexit",
     "test_ifsck",
     "test_igroupadmin",
+    "test_ihelp",
+    "test_ilocate",
     "test_ils",
     "test_ilsresc",
     "test_imeta_error_handling",
@@ -38,6 +42,7 @@
     "test_irule",
     "test_iscan",
     "test_istream",
+    "test_isysmeta",
     "test_iticket",
     "test_itouch",
     "test_itree",
@@ -49,8 +54,8 @@
     "test_misc",
     "test_negotiation",
     "test_native_rule_engine_plugin",
-    "test_python_rule_engine_plugin",
     "test_prep_genquery_iterator",
+    "test_python_rule_engine_plugin",
     "test_quotas",
     "test_resource_configuration",
     "test_resource_types.Test_Resource_Compound",
